dockerfiles/scripts/requirements.txt
SQLAlchemy==2.0.39
python-dotenv==1.0.1
PyMySQL==1.1.1